# CUDA: faster FlashAttention for batch sizes > 1