@kalomaze 2024-04-13
Maybe there's something I'm missing here, but doesn't the mention of ~100t/s prompt processing for batch size 1 seem inaccurate? I think you should be able to hit 1000t/s on quantized 7b for bs1, unless the actual token batch count is set to something low like 64 or something?