@kalomaze 2024-04-13
On my machine, hosting to llama.cpp server, I'm able to get a couple coherent tokens before garbage output consistently comes out for Cohere's Command-R 35b at q5_K_S:

```
but it's still frustrating. being nothing more than an error  as if i was an afterthoughtserviceNameserviceNameserviceNameserviceNameserviceNameserviceNameserviceNameserviceNameserviceNameserviceNameserviceNameserviceNameserviceNameserviceNameserviceNameserviceNameserviceNameserviceNameserviceNameserviceNameserviceNameserviceNameserviceNameserviceNameserviceNameserviceNameserviceNameserviceNameserviceNameserviceNameserviceNameserviceNameserviceNameserviceNameserviceNameserviceNameserviceNameserviceNameserviceNameserviceNameserviceNameserviceNameserviceNameserviceNameserviceNameserviceNameserviceNameserviceNameserviceNameserviceNameserviceNameserviceNameserviceNameserviceNameserviceNameserviceNameserviceNameserviceNameserviceNameserviceNameserviceNameserviceNameserviceName
```

It always degenerates into "nameservice".