@JohannesGaessler 2024-04-13
>Maybe there's something I'm missing here, but doesn't the mention of ~100t/s prompt processing for batch size 1 seem inaccurate? I think you should be able to hit 1000t/s on q4_0 7b for bs1 on master easy with a 3090, unless the actual token batch count is what is described here and it's not about concurrent requests?

>I think I'm just mixing up my definitions of "batching" here and it's not about concurrent requests.

The numbers are from running something like

```
export model_name=llama_2-7b && export quantization=q4_0
./llama-bench --model models/opt/${model_name}-${quantization}.gguf -n 0 -p 4096 -r 1 -b 1,2,4,8,16,32,64,128,256,512,1024,2048,4096
```

So they are representative of a single concurrent context with a varying number of tokens processed in parallel. You would essentially get these speeds for prompt processing if you were to set this batch size. batch size 1 is equivalent to generation.

>On my machine, hosting to llama.cpp server, I'm able to get a couple coherent tokens, sometimes up to a short paragraph in length, before garbage output consistently comes out for Cohere's Command-R 35b at q5_K_S:

I can reproduce the issue. It's a bug in the batch size 1 kernel that is already on `gg/flash-attn`.