@JohannesGaessler 2024-04-13
The problem seemed to be that there was a race condition that - depending on the specifics - could sometimes result in garbage outputs. @kalomaze can you confirm that the fix works?