@kalomaze 2024-04-14
> The problem seemed to be that there was a race condition that - depending on the specifics - could sometimes result in garbage outputs. @kalomaze can you confirm that the fix works?

Seems fine on my end now! Responses are perfectly normal on regen / fresh context ingest / etc.

I ran a full offload of a q6_K 20b model (not a new / noteworthy one, just was convenient) to test out different `nvidia-smi` power limits.

GPU      | Model              | Batch size | Test   | t/s master | t/s jg/flash-attn-20 | Speedup
----------|-------------------|------------|--------|------------|----------------------|--------
RTX 3090 (250w) | internlm 20B Q6_K | 512        | pp4096 | 873        | 1112                 | 1.27
RTX 3090 (300w) | internlm 20B Q6_K | 512        | pp4096 | 1159        | 1403                 | 1.21

Seems great @JohannesGaessler. Is there anything else left for FlashAttention to be merged into mainline