@phymbert 2024-04-14
> 📈 **llama.cpp server** for _bench-server-baseline_ on _Standard_NC4as_T4_v3_ for `phi-2`-`q4_0`: **272 iterations** 🚀

@JohannesGaessler Thanks for the work here, I am out of the flash attention branches, but I would like to understand why the bench workflow is slower, 272 iterations here versus [426](https://github.com/ggerganov/llama.cpp/pull/6635#issuecomment-2054020506) on master. It runs on phi-2 Q4_0. Thanks