@JohannesGaessler 2024-04-14
>Seems great @JohannesGaessler. Is there anything else left for FlashAttention to be merged into mainline

You'll have to ask @ggerganov about that.

>@JohannesGaessler Thanks for the work here, I am out of the flash attention branches, but I would like to understand why the bench workflow is slower, 272 iterations here versus https://github.com/ggerganov/llama.cpp/pull/6635#issuecomment-2054020506 on master. It runs on phi-2 Q4_0. Thanks

For this PR the bot reports 241.82 t/s pp and 42.58 t/s tg. In the post you linked it's 118.38 t/s pp and 24.06 t/s tg. Aren't those the relevant numbers?