@phymbert 2024-04-14
> For this PR the bot reports 241.82 t/s pp and 42.58 t/s tg. In the post you linked it's 118.38 t/s pp and 24.06 t/s tg. Aren't those the relevant numbers?

Yes you are right thanks. It's a lot faster per sequence. Then I will investigate why the server is "globally" slower with 8 concurrent sequences/slots/users. @ggerganov I think we need to fix that before merging.