@phymbert 2024-04-14
> For this PR the bot reports 241.82 t/s pp and 42.58 t/s tg. In the post you linked it's 118.38 t/s pp and 24.06 t/s
> tg. Aren't those the relevant numbers?

No, the bench has been fixed since, it is now using the metrics from the server side, not the client side.
If you look at the PP graph (which is built by the server + prometheus), the prompt processing is twice slower:

| Branch                                                                                                        | Iterations | PP                                                                                                    | TG                                                                                                      |
|---------------------------------------------------------------------------------------------------------------|------------|-------------------------------------------------------------------------------------------------------|---------------------------------------------------------------------------------------------------------|
| [master](https://github.com/ggerganov/llama.cpp/pull/6635#issuecomment-2054020506)                            | 426        | <img width="100%" height="100%" src="https://i.imgur.com/2SfSwt5.jpeg" alt="prompt_tokens_seconds" />   | <img width="100%" height="100%" src="https://i.imgur.com/JLibKko.jpeg" alt="predicted_tokens_seconds"/>   |
| [gg/flash-attn](https://github.com/ggerganov/llama.cpp/pull/5021#issuecomment-2025797454)                     | 264        | <img width="100%" height="100%" src="https://i.imgur.com/w50UpQJ.jpeg" alt="prompt_tokens_seconds" /> | <img width="100%" height="100%" src="https://i.imgur.com/P9DyMRw.jpeg" alt="predicted_tokens_seconds"/> |
| [JohannesGaessler:jg/flash-attn-20](https://github.com/ggerganov/llama.cpp/pull/6646#issuecomment-2053980311) | 272        | <img width="100%" height="100%" src="https://i.imgur.com/wSH5UPQ.jpeg" alt="prompt_tokens_seconds" /> | <img width="100%" height="100%" src="https://i.imgur.com/2Wc7mAJ.jpeg" alt="predicted_tokens_seconds"/> |