@JohannesGaessler 2024-04-14
I don't have a comprehensive overview of the changes in gg/flash-attn. And I don't know the details of how the server benchmark presented here works either. **If** the issue is due to the CUDA kernels and not some other changes it could just be that I tested and optimized the code on Ampere and Ada Lovelace and that for whatever reason the performance is just bad on Turing. It could also be that it's an issue specific to Phi-2 since that particular model has a janky head size of 80, If you have access to T4s, do a quick benchmark on them to check whether the issue is specific to T4s, to the server, or to Phi-2. Or tell me how I can run this benchmark locally on my own hardware.